Add ALL

+++ app/Http/Controllers/Assignments/UserController.php
@@ -242,7 +242,7 @@ public function push(Request $request)
                                 //$classpath = file_exists($destinationPath.$filename.'.class');
                                 unlink($checkpath);
                                 if (strpos($errorpath, '\u') !== false) {
-                                    dd("come to compile normally");
+                                  //  dd("come to compile normally");
                                    $executeq = 'javac -d '.$destinationPath.' '.$final.' 2> '.$destinationPath.'error-'.$filename;
                                   //  print_r($executeq);
                                     $result = shell_exec($executeq);    
@@ -250,10 +250,13 @@ public function push(Request $request)
                                    // dd($checkpath);
                                     $errorpath = File::get($checkpath);
                                     $classpath = file_exists($destinationPath.$filename.'.class');
+                                    if(!empty($errorpath) && $classpath == 0)
+                                    {
                                     unlink($checkpath);
                                     $tmper = str_replace("//",  "/", $destinationPath);
                                     $showme = str_replace($tmper, "Compiler:", $errorpath);
                                     return view('assignments.error',['errorpath'=>$showme]);
+                                    }
                                 }
                                 $tmper = str_replace("//",  "/", $destinationPath);
                                 $showme = str_replace($tmper, "Compiler:", $errorpath);
